mgr/rgw: initial work
Basic catch all command that passes through params to radosgw-admin
